drm/verisilicon: unconditionally use the DC device as dma_dev
As the display-subsystem is outside of the soc device tree node (where
the dma-noncoherent property is added), it's assumed to be coherent,
and lead to allocated memory being not properly mapped (still cached).

Unconditionally use the DC device (which is in the soc node) as dma_dev
to fix framebuffer mapping issue.

Signed-off-by: Icenowy Zheng <uwu@icenowy.me>